# openapi/guzo-v1.json
{
  "components": {
    "schemas": {
...
      "Currency": {
        "enum": [
          "ETB",
          "USD"
        ],
        "title": "Currency",
        "type": "string"
      },
      "DriverResponse": {
        "properties": {
          "driver": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "vehicles": {
            "items": {
              "$ref": "#/components/schemas/VehicleResponse"
            },
            "title": "Vehicles",
            "type": "array"
          }
        },
        "required": [
          "driver",
          "vehicles"
        ],
        "title": "DriverResponse",
        "type": "object"
      },
      "DriverUpdate": {
        "properties": {
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "photo_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          }
        },
        "title": "DriverUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Money": {
        "description": "An amount in minor units (santim, cents). Never a float.",
        "properties": {
          "amount_minor": {
            "title": "Amount Minor",
            "type": "integer"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          }
        },
        "required": [
          "amount_minor",
          "currency"
        ],
        "title": "Money",
        "type": "object"
      },
...
      "PartnerAttribution": {
        "description": "What a partner's code has brought in. Commission is owed on completed trips only.",
        "properties": {
          "bookings_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Bookings By Status",
            "type": "object"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "commission_due": {
            "items": {
              "$ref": "#/components/schemas/Money"
            },
            "title": "Commission Due",
            "type": "array"
          },
          "completed_fares": {
            "items": {
              "$ref": "#/components/schemas/Money"
            },
            "title": "Completed Fares",
            "type": "array"
          }
        },
        "required": [
          "code",
          "bookings_by_status",
          "completed_fares",
          "commission_due"
        ],
        "title": "PartnerAttribution",
        "type": "object"
      },
      "PartnerCreate": {
        "properties": {
          "code": {
            "pattern": "^[A-Za-z0-9_-]{3,40}$",
            "title": "Code",
            "type": "string"
          },
          "commission_pct": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Commission Pct",
            "type": "integer"
          },
          "kind": {
            "$ref": "#/components/schemas/PartnerKind"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "code",
          "kind",
          "name",
          "commission_pct"
        ],
        "title": "PartnerCreate",
        "type": "object"
      },
      "PartnerKind": {
        "enum": [
          "hotel",
          "host",
          "agency"
        ],
        "title": "PartnerKind",
        "type": "string"
      },
      "PartnerPublic": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/PartnerKind"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "code",
          "kind",
          "name"
        ],
        "title": "PartnerPublic",
        "type": "object"
      },
      "PartnerResponse": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "commission_pct": {
            "title": "Commission Pct",
            "type": "integer"
          },
          "kind": {
            "$ref": "#/components/schemas/PartnerKind"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "code",
          "kind",
          "name",
          "commission_pct",
          "active"
        ],
        "title": "PartnerResponse",
        "type": "object"
      },
      "PartnerUpdate": {
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "commission_pct": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commission Pct"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "PartnerUpdate",
        "type": "object"
      },
...
      "Role": {
        "enum": [
          "booker",
          "driver",
          "ops"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "UserResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_verified": {
            "title": "Is Verified",
            "type": "boolean"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "total_ratings": {
            "title": "Total Ratings",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "role",
          "phone",
          "email",
          "name",
          "photo_url",
          "is_verified",
          "rating",
          "total_ratings",
          "created_at"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VehicleClass": {
        "enum": [
          "sedan",
          "minivan",
          "suv"
        ],
        "title": "VehicleClass",
        "type": "string"
      },
      "VehicleCreate": {
        "properties": {
          "color": {
            "maxLength": 30,
            "minLength": 1,
            "title": "Color",
            "type": "string"
          },
          "make": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Make",
            "type": "string"
          },
          "model": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "plate": {
            "maxLength": 20,
            "minLength": 2,
            "title": "Plate",
            "type": "string"
          },
          "vehicle_class": {
            "$ref": "#/components/schemas/VehicleClass"
          }
        },
        "required": [
          "plate",
          "make",
          "model",
          "color",
          "vehicle_class"
        ],
        "title": "VehicleCreate",
        "type": "object"
      },
...
      "VehicleResponse": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "color": {
            "maxLength": 30,
            "minLength": 1,
            "title": "Color",
            "type": "string"
          },
          "driver_id": {
            "title": "Driver Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "make": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Make",
            "type": "string"
          },
          "model": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "plate": {
            "maxLength": 20,
            "minLength": 2,
            "title": "Plate",
            "type": "string"
          },
          "vehicle_class": {
            "$ref": "#/components/schemas/VehicleClass"
          }
        },
        "required": [
          "plate",
          "make",
          "model",
          "color",
          "vehicle_class",
          "id",
          "driver_id",
          "active"
        ],
        "title": "VehicleResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Guzo API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/ops/drivers": {
      "get": {
        "operationId": "list_drivers",
        "parameters": [
          {
            "in": "query",
            "name": "verified",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Verified"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DriverResponse"
                  },
                  "title": "Response List Drivers",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Drivers",
        "tags": [
          "ops"
        ]
      }
    },
    "/v1/ops/drivers/{driver_id}": {
      "patch": {
        "operationId": "update_driver",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "title": "Driver Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Driver",
        "tags": [
          "ops"
        ]
      }
    },
    "/v1/ops/drivers/{driver_id}/vehicles": {
      "post": {
        "operationId": "add_vehicle",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "title": "Driver Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Vehicle",
        "tags": [
          "ops"
        ]
      }
    },
    "/v1/ops/drivers/{driver_id}/vehicles/{vehicle_id}": {
      "delete": {
        "description": "Take a vehicle out of service. It stays on past bookings.",
        "operationId": "retire_vehicle",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "title": "Driver Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "vehicle_id",
            "required": true,
            "schema": {
              "title": "Vehicle Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retire Vehicle",
        "tags": [
          "ops"
        ]
      }
    },
    "/v1/ops/drivers/{driver_id}/verify": {
      "post": {
        "description": "Mark a driver as vetted. Only verified drivers can be assigned.",
        "operationId": "verify_driver",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "title": "Driver Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Verify Driver",
        "tags": [
          "ops"
        ]
      }
    },
    "/v1/ops/partners": {
      "get": {
        "operationId": "list_partners",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PartnerResponse"
                  },
                  "title": "Response List Partners",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Partners",
        "tags": [
          "partners",
          "ops"
        ]
      },
      "post": {
        "operationId": "create_partner",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Partner",
        "tags": [
          "partners",
          "ops"
        ]
      }
    },
    "/v1/ops/partners/{code}": {
      "patch": {
        "operationId": "update_partner",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Partner",
        "tags": [
          "partners",
          "ops"
        ]
      }
    },
    "/v1/ops/partners/{code}/attribution": {
      "get": {
        "operationId": "partner_attribution",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerAttribution"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Partner Attribution",
        "tags": [
          "partners",
          "ops"
        ]
      }
    },
    "/v1/partners/{code}": {
      "get": {
        "description": "Lets a booking page check a referral code and show who it belongs to.",
        "operationId": "get_partner",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Partner",
        "tags": [
          "partners"
        ]
      }
    },
...
  }
}